Move map to pulse
